+++ R/miscellanous.R
@@ -126,3 +126,53 @@ simplify_taxo <- function(physeq) {
   physeq@tax_table <- taxo
   return(physeq)
 }
+
+################################################################################
+#' Get the extension of a file
+#' 
+#' @param file (required): path to a file
+#' @description
+#' `r lifecycle::badge("maturing")`
+#'
+#' @author Adrien Taudière
+#'
+#' @return A  \code{\link{phyloseq-class}} object with simplified taxonomy
+#' @export
+get_file_extension <- function(file){ 
+  file_ext <- strsplit(basename(file), ".", fixed = TRUE)[[1]][-1]
+  return(file_ext)
+}
+
+
+#' Count sequences 
+#'
+#' @description
+#'  `r lifecycle::badge("experimental")`
+#'   Use grep to count the number of line with only one '+' (fastq, fastq.gz) 
+#'   or lines starting with a '>' (fasta) to count sequences.
+#'
+#' @param file (required) The path to a  fasta, fastq or fastq.gz file
+#'
+#' @return the number of sequences
+#' @author Adrien Taudière
+#' @export
+#' 
+count_seq <- function(file = NULL) {
+
+  if(get_file_extension(file) %in% "fasta"){
+    seq_nb <- system(paste0("cat " , file, " | grep -ce '^>'"),
+                   intern = TRUE)
+  } else if (get_file_extension(file) %in% "fastq"){
+    if(get_file_extension(file) %in% "gz"){
+      seq_nb <- system(paste0("zcat " , file, " | grep -ce '^+$'"),
+                     intern = TRUE)
+    }
+    else {
+      seq_nb <- system(paste0("cat " , file, " | grep -ce '^+$'"),
+                   intern = TRUE)
+    }
+  } else {
+    stop(paste0("The file extension",get_file_extension(file) ,"") 
+  }
+  return(as.numeric(seq_nb))
+}