+++ R/targets_misc.R
@@ -47,7 +47,8 @@ MM_idtaxa <- function(test, trainingSet, column_names = c("Kingdom_idtaxa", "Phy
 #'
 #' @return a list of one (single end) or two (paired end) list of files
 #' @export
-#'
+#' @author Adrien Taudière
+
 list_fastq_files <-
   function(path,
            paired_end = TRUE,
@@ -74,6 +75,40 @@ list_fastq_files <-
   }
 
 
+#' Count sequences 
+#'
+#' @description
+#'  `r lifecycle::badge("experimental")`
+#'   Use grep to count the number of line with only one '+' (fastq, fastq.gz) 
+#'   or lines starting with a '>' (fasta) to count sequences.
+#'
+#' @param file (required) The path to a  fasta, fastq or fastq.gz file
+#'
+#' @return the number of sequences
+#' @author Adrien Taudière
+#' @export
+#' 
+count_seq <- function(file = NULL) {
+
+  if(get_file_extension(file) %in% "fasta"){
+    seq_nb <- system(paste0("cat " , file, " | grep -ce '^>'"),
+                   intern = TRUE)
+  } else if (get_file_extension(file) %in% "fastq"){
+    if(get_file_extension(file) %in% "gz"){
+      seq_nb <- system(paste0("zcat " , file, " | grep -ce '^+$'"),
+                     intern = TRUE)
+    }
+    else {
+      seq_nb <- system(paste0("cat " , file, " | grep -ce '^+$'"),
+                   intern = TRUE)
+    }
+  } else {
+    stop(paste0("The file extension",get_file_extension(file) ,"") 
+  }
+  return(as.numeric(seq_nb))
+}
+
+
 #' Rename samples of an otu_table
 #'
 #' @description
@@ -89,6 +124,8 @@ list_fastq_files <-
 #'
 #' @export
 #' @md
+#' 
+#' @author Adrien Taudière
 #'
 #' @examples
 #' data(data_fungi)